# microservices/integration_service/docs/API_CONTRACTS.yaml
openapi: 3.0.3
info:
  title: Integration Service API
  description: |
    UK Management Bot - Integration Service API

    Manages external integrations including:
    - Google Sheets
    - Geocoding (Google Maps, Yandex Maps)
    - Webhooks
    - Custom integrations

    Features:
    - Multi-tenant isolation via X-Management-Company-Id header
    - Rate limiting and quota management
    - Response caching
    - Event-driven architecture
    - Comprehensive logging and monitoring
  version: 1.0.0
  contact:
    name: UK Management Bot Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8006/api/v1
    description: Local development
  - url: https://integration.uk-management.com/api/v1
    description: Production
security:
  - BearerAuth: []
  - ServiceToken: []
...
paths:
...
  /sheets/read:
    post:
      tags: [google-sheets]
      summary: Read from Google Sheet
      description: Read data from specified Google Sheet range
      parameters:
        - $ref: '#/components/parameters/ManagementCompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spreadsheet_id, range]
              properties:
                spreadsheet_id:
                  type: string
                  description: Google Sheet ID
                range:
                  type: string
                  description: A1 notation range (e.g., "Sheet1!A1:D10")
                value_render_option:
                  type: string
                  enum: [FORMATTED_VALUE, UNFORMATTED_VALUE, FORMULA]
                  default: FORMATTED_VALUE
      responses:
        '200':
          description: Sheet data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  values:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
                  range:
                    type: string
                  cached:
                    type: boolean
  /sheets/write:
    post:
      tags: [google-sheets]
      summary: Write to Google Sheet
      description: Write data to specified Google Sheet range
      parameters:
        - $ref: '#/components/parameters/ManagementCompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spreadsheet_id, range, values]
              properties:
                spreadsheet_id:
                  type: string
                range:
                  type: string
                values:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
                value_input_option:
                  type: string
                  enum: [RAW, USER_ENTERED]
                  default: USER_ENTERED
      responses:
        '200':
          description: Data written successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_range:
                    type: string
                  updated_rows:
                    type: integer
                  updated_columns:
                    type: integer
  /sheets/append:
    post:
      tags: [google-sheets]
      summary: Append to Google Sheet
      description: Append data to the end of a Google Sheet
      parameters:
        - $ref: '#/components/parameters/ManagementCompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spreadsheet_id, range, values]
              properties:
                spreadsheet_id:
                  type: string
                range:
                  type: string
                values:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
      responses:
        '200':
          description: Data appended successfully

  # ============ GEOCODING ============
...
  /webhooks:
    get:
      tags: [webhooks]
      summary: List webhooks
      parameters:
        - $ref: '#/components/parameters/ManagementCompanyId'
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookConfig'

    post:
      tags: [webhooks]
      summary: Create webhook
      parameters:
        - $ref: '#/components/parameters/ManagementCompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookConfigCreate'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookConfig'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ServiceToken:
      type: apiKey
      in: header
      name: X-Service-Token
  parameters:
    ManagementCompanyId:
      name: X-Management-Company-Id
      in: header
      required: true
      schema:
        type: string
      description: Management Company ID for multi-tenancy
...
  schemas:
...
    WebhookConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        webhook_name:
          type: string
        webhook_url:
          type: string
        source_service:
          type: string
        event_types:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        total_received:
          type: integer
        total_successful:
          type: integer
        total_failed:
          type: integer
    WebhookConfigCreate:
      type: object
      required: [webhook_name, source_service, event_types]
      properties:
        webhook_name:
          type: string
        source_service:
          type: string
        event_types:
          type: array
          items:
            type: string
        secret_key:
          type: string
...